created weatherReducer.js, reducer dir and action dir
